Develop (#2)
* Make it work with old ff.net

With this, it basically works with the old (pre-2002) version of the website by
abusing the CDX API to get a list of everything that war archived.

It's not optimized yet but you can grab content with it. Nice.

* Deal with invalid character errors

Ignore the response for now if it contains any. Not the best solution.

* Update README

* Refactor code