Develop (#3)
* Make it work with old ff.net

With this, it basically works with the old (pre-2002) version of the website by
abusing the CDX API to get a list of everything that war archived.

It's not optimized yet but you can grab content with it. Nice.

* Deal with invalid character errors

Ignore the response for now if it contains any. Not the best solution.

* Update README

* Refactor code

* Optimize old scraper

* Add first version of new scraper

This isn't really working yet, but we're getting there.

* Finalize first version of scraper for new ffnet

This version is able to download an archived story given a memento URI and a
timestamp, complete with extracting metadata. Not tested yet.

* Add feature to delete empty folders after download failure

* Add a memento struct that holds memento URIs and timestamps

This streamlines the CDX responses for both new and old scrapers.