Improve new scraper
Adds some additional checks before attempting to fetch the story (e.g. if the
story is actually inside the response, b/c it could be a "story not found" page
that has been archived) and whether title and categories could be extracted.

It also adds a way of telling the scraper to only fetch a few stories because
the new format does not group them, so the CDX response may countain tens of
thousands of stories.

Also we cache the response now, which makes this ten times faster.