Develop (#4)
* Make it work with old ff.net

With this, it basically works with the old (pre-2002) version of the website by
abusing the CDX API to get a list of everything that war archived.

It's not optimized yet but you can grab content with it. Nice.

* Deal with invalid character errors

Ignore the response for now if it contains any. Not the best solution.

* Update README

* Refactor code

* Optimize old scraper

* Add first version of new scraper

This isn't really working yet, but we're getting there.

* Finalize first version of scraper for new ffnet

This version is able to download an archived story given a memento URI and a
timestamp, complete with extracting metadata. Not tested yet.

* Add feature to delete empty folders after download failure

* Add a memento struct that holds memento URIs and timestamps

This streamlines the CDX responses for both new and old scrapers.

* Fix a type error with the memento struct

Tested-by: Alex Sun <alexsun82@icloud.com>

* Improve error output for old scraper

Adds a newline at the end of an error response. :D

* Improve new scraper

Adds some additional checks before attempting to fetch the story (e.g. if the
story is actually inside the response, b/c it could be a "story not found" page
that has been archived) and whether title and categories could be extracted.

It also adds a way of telling the scraper to only fetch a few stories because
the new format does not group them, so the CDX response may countain tens of
thousands of stories.

Also we cache the response now, which makes this ten times faster.

* Output more information

* Update readme